silhouette_maker/requirements.txt
Pillow
webuiapi
